INS-499: move tests to different files
