+++ functest/main_test.go
@@ -18,13 +18,9 @@ package functest
 
 import (
 	"bufio"
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"go/build"
 	"io"
-	"io/ioutil"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -32,11 +28,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/insolar/insolar/api"
 	"github.com/insolar/insolar/logicrunner/goplugin/testutil"
 	"github.com/insolar/insolar/testutils"
 	"github.com/pkg/errors"
-	"github.com/stretchr/testify/assert"
 )
 
 const HOST = "http://localhost:19191"
@@ -270,408 +264,4 @@ func testMainWrapper(m *testing.M) int {
 
 func TestMain(m *testing.M) {
 	os.Exit(testMainWrapper(m))
-}
-
-type postParams map[string]interface{}
-
-type errorResponse struct {
-	Code    int    `json:"code"`
-	Message string `json:"message"`
-}
-
-type responseInterface interface {
-	getError() *errorResponse
-}
-
-type baseResponse struct {
-	Qid string         `json:"qid"`
-	Err *errorResponse `json:"error"`
-}
-
-func (r *baseResponse) getError() *errorResponse {
-	return r.Err
-}
-
-type createMemberResponse struct {
-	baseResponse
-	Reference string `json:"reference"`
-}
-
-type sendMoneyResponse struct {
-	baseResponse
-	Success bool `json:"success"`
-}
-
-type getBalanceResponse struct {
-	baseResponse
-	Amount   uint   `json:"amount"`
-	Currency string `json:"currency"`
-}
-
-type userInfo struct {
-	baseResponse
-	Member string `json:"member"`
-	Wallet uint   `json:"wallet"`
-}
-
-type dumpAllUsersResponse struct {
-	baseResponse
-	DumpInfo []userInfo `json:"dump_info"`
-}
-
-func createMember(t *testing.T) string {
-	body := getResponseBody(t, postParams{
-		"query_type": "create_member",
-		"name":       testutils.RandomString(),
-	})
-
-	firstMemberResponse := &createMemberResponse{}
-	unmarshalResponse(t, body, firstMemberResponse)
-
-	return firstMemberResponse.Reference
-}
-
-func getBalance(t *testing.T, reference string) int {
-	body := getResponseBody(t, postParams{
-		"query_type": "get_balance",
-		"reference":  reference,
-	})
-
-	firstBalanceResponse := &getBalanceResponse{}
-	unmarshalResponse(t, body, firstBalanceResponse)
-
-	return int(firstBalanceResponse.Amount)
-}
-
-func getResponseBody(t *testing.T, postParams map[string]interface{}) []byte {
-	jsonValue, _ := json.Marshal(postParams)
-	postResp, err := http.Post(TestURL, "application/json", bytes.NewBuffer(jsonValue))
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, postResp.StatusCode)
-	body, err := ioutil.ReadAll(postResp.Body)
-	assert.NoError(t, err)
-	return body
-}
-
-func unmarshalResponse(t *testing.T, body []byte, response responseInterface) {
-	err := json.Unmarshal(body, &response)
-	assert.NoError(t, err)
-	assert.Nil(t, response.getError())
-}
-
-func unmarshalResponseWithError(t *testing.T, body []byte, response responseInterface) {
-	err := json.Unmarshal(body, &response)
-	assert.NoError(t, err)
-	assert.NotNil(t, response.getError())
-}
-
-func TestInsolardResponseNotErr(t *testing.T) {
-	body := getResponseBody(t, postParams{
-		"query_type": "dump_all_users",
-	})
-
-	response := &dumpAllUsersResponse{}
-	unmarshalResponse(t, body, response)
-
-	assert.Nil(t, response.Err)
-}
-
-func TestTransferMoney(t *testing.T) {
-	firstMemberRef := createMember(t)
-	secondMemberRef := createMember(t)
-	oldFirstBalance := getBalance(t, firstMemberRef)
-	oldSecondBalance := getBalance(t, secondMemberRef)
-
-	amount := 111
-
-	// Transfer money from one member to another
-	body := getResponseBody(t, postParams{
-		"query_type": "send_money",
-		"from":       secondMemberRef,
-		"to":         firstMemberRef,
-		"amount":     amount,
-	})
-
-	transferResponse := &sendMoneyResponse{}
-	unmarshalResponse(t, body, transferResponse)
-
-	assert.Equal(t, true, transferResponse.Success)
-
-	newFirstBalance := getBalance(t, firstMemberRef)
-	newSecondBalance := getBalance(t, secondMemberRef)
-
-	assert.Equal(t, oldFirstBalance+amount, newFirstBalance)
-	assert.Equal(t, oldSecondBalance-amount, newSecondBalance)
-}
-
-func TestWrongUrl(t *testing.T) {
-	jsonValue, _ := json.Marshal(postParams{
-		"query_type": "dump_all_users",
-	})
-	testURL := HOST + "/not_api/v1"
-	postResp, err := http.Post(testURL, "application/json", bytes.NewBuffer(jsonValue))
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusNotFound, postResp.StatusCode)
-}
-
-func TestGetRequest(t *testing.T) {
-	postResp, err := http.Get(TestURL)
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, postResp.StatusCode)
-	body, err := ioutil.ReadAll(postResp.Body)
-	assert.NoError(t, err)
-
-	getResponse := &baseResponse{}
-	unmarshalResponseWithError(t, body, getResponse)
-
-	assert.Equal(t, api.BadRequest, getResponse.Err.Code)
-	assert.Equal(t, "Bad request", getResponse.Err.Message)
-}
-
-func TestWrongJson(t *testing.T) {
-	postResp, err := http.Post(TestURL, "application/json", bytes.NewBuffer([]byte("some not json value")))
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, postResp.StatusCode)
-	body, err := ioutil.ReadAll(postResp.Body)
-	assert.NoError(t, err)
-
-	response := &baseResponse{}
-	unmarshalResponseWithError(t, body, response)
-
-	assert.Equal(t, api.BadRequest, response.Err.Code)
-	assert.Equal(t, "Bad request", response.Err.Message)
-}
-
-func TestWrongQueryType(t *testing.T) {
-	body := getResponseBody(t, postParams{
-		"query_type": "wrong_query_type",
-	})
-
-	response := &baseResponse{}
-	unmarshalResponseWithError(t, body, response)
-
-	assert.Equal(t, api.BadRequest, response.Err.Code)
-	assert.Equal(t, "Wrong query parameter 'query_type' = 'wrong_query_type'", response.Err.Message)
-}
-
-func TestWithoutQueryType(t *testing.T) {
-	body := getResponseBody(t, postParams{})
-
-	response := &baseResponse{}
-	unmarshalResponseWithError(t, body, response)
-
-	assert.Equal(t, api.BadRequest, response.Err.Code)
-	assert.Equal(t, "Wrong query parameter 'query_type' = ''", response.Err.Message)
-}
-
-func TestTooMuchParams(t *testing.T) {
-	body := getResponseBody(t, postParams{
-		"query_type": "create_member",
-		"some_param": "irrelevant info",
-		"name":       testutils.RandomString(),
-	})
-
-	firstMemberResponse := &createMemberResponse{}
-	unmarshalResponse(t, body, firstMemberResponse)
-
-	firstMemberRef := firstMemberResponse.Reference
-	assert.NotEqual(t, "", firstMemberRef)
-}
-
-func TestQueryTypeAsIntParams(t *testing.T) {
-	body := getResponseBody(t, postParams{
-		"query_type": 100,
-	})
-
-	response := &baseResponse{}
-	unmarshalResponseWithError(t, body, response)
-
-	assert.Equal(t, api.BadRequest, response.Err.Code)
-	assert.Equal(t, "Bad request", response.Err.Message)
-}
-
-func TestWrongTypeInParams(t *testing.T) {
-	body := getResponseBody(t, postParams{
-		"query_type": "create_member",
-		"name":       128182187,
-	})
-
-	response := &baseResponse{}
-	unmarshalResponseWithError(t, body, response)
-
-	assert.Equal(t, api.BadRequest, response.Err.Code)
-	assert.Equal(t, "Bad request", response.Err.Message)
-}
-
-// TODO: unskip test after doing errors in smart contracts
-func _TestWrongReferenceInParams(t *testing.T) {
-	body := getResponseBody(t, postParams{
-		"query_type": "get_balance",
-		"reference":  testutils.RandomString(),
-	})
-
-	response := &baseResponse{}
-	unmarshalResponseWithError(t, body, response)
-
-	assert.Equal(t, api.BadRequest, response.Err.Code)
-	assert.Equal(t, "Bad request", response.Err.Message)
-}
-
-func TestTransferNegativeAmount(t *testing.T) {
-	firstMemberRef := createMember(t)
-	secondMemberRef := createMember(t)
-	oldFirstBalance := getBalance(t, firstMemberRef)
-	oldSecondBalance := getBalance(t, secondMemberRef)
-
-	body := getResponseBody(t, postParams{
-		"query_type": "send_money",
-		"from":       secondMemberRef,
-		"to":         firstMemberRef,
-		"amount":     -111,
-	})
-
-	transferResponse := &sendMoneyResponse{}
-	unmarshalResponseWithError(t, body, transferResponse)
-
-	assert.Equal(t, api.BadRequest, transferResponse.Err.Code)
-	assert.Equal(t, "Bad request", transferResponse.Err.Message)
-
-	newFirstBalance := getBalance(t, firstMemberRef)
-	newSecondBalance := getBalance(t, secondMemberRef)
-
-	assert.Equal(t, oldFirstBalance, newFirstBalance)
-	assert.Equal(t, oldSecondBalance, newSecondBalance)
-
-}
-
-func TestTransferAllAmount(t *testing.T) {
-	firstMemberRef := createMember(t)
-	secondMemberRef := createMember(t)
-	oldFirstBalance := getBalance(t, firstMemberRef)
-	oldSecondBalance := getBalance(t, secondMemberRef)
-
-	body := getResponseBody(t, postParams{
-		"query_type": "send_money",
-		"from":       secondMemberRef,
-		"to":         firstMemberRef,
-		"amount":     oldSecondBalance,
-	})
-
-	transferResponse := &sendMoneyResponse{}
-	unmarshalResponse(t, body, transferResponse)
-
-	assert.Equal(t, true, transferResponse.Success)
-
-	newFirstBalance := getBalance(t, firstMemberRef)
-	newSecondBalance := getBalance(t, secondMemberRef)
-
-	assert.Equal(t, oldFirstBalance+oldSecondBalance, newFirstBalance)
-	assert.Equal(t, 0, newSecondBalance)
-
-}
-
-func _TestTransferMoreThanAvailableAmount(t *testing.T) {
-	firstMemberRef := createMember(t)
-	secondMemberRef := createMember(t)
-	oldFirstBalance := getBalance(t, firstMemberRef)
-	oldSecondBalance := getBalance(t, secondMemberRef)
-
-	body := getResponseBody(t, postParams{
-		"query_type": "send_money",
-		"from":       secondMemberRef,
-		"to":         firstMemberRef,
-		"amount":     10000000000,
-	})
-
-	transferResponse := &sendMoneyResponse{}
-	unmarshalResponse(t, body, transferResponse)
-
-	// Add checking than contract gives specific error
-
-	newFirstBalance := getBalance(t, firstMemberRef)
-	newSecondBalance := getBalance(t, secondMemberRef)
-
-	assert.Equal(t, oldFirstBalance, newFirstBalance)
-	assert.Equal(t, oldSecondBalance, newSecondBalance)
-}
-
-func _TestTransferToMyself(t *testing.T) {
-	memberRef := createMember(t)
-	oldBalance := getBalance(t, memberRef)
-
-	body := getResponseBody(t, postParams{
-		"query_type": "send_money",
-		"from":       memberRef,
-		"to":         memberRef,
-		"amount":     oldBalance - 1,
-	})
-
-	transferResponse := &sendMoneyResponse{}
-	unmarshalResponse(t, body, transferResponse)
-
-	assert.Equal(t, true, transferResponse.Success)
-
-	newBalance := getBalance(t, memberRef)
-
-	assert.Equal(t, oldBalance, newBalance)
-}
-
-// TODO: test to check overflow of balance
-// TODO: check transfer zero amount
-
-func TestTransferTwoTimes(t *testing.T) {
-	firstMemberRef := createMember(t)
-	secondMemberRef := createMember(t)
-	oldFirstBalance := getBalance(t, firstMemberRef)
-	oldSecondBalance := getBalance(t, secondMemberRef)
-
-	firstBody := getResponseBody(t, postParams{
-		"query_type": "send_money",
-		"from":       secondMemberRef,
-		"to":         firstMemberRef,
-		"amount":     100,
-	})
-	transferResponse := &sendMoneyResponse{}
-	unmarshalResponse(t, firstBody, transferResponse)
-	assert.Equal(t, true, transferResponse.Success)
-
-	secondBody := getResponseBody(t, postParams{
-		"query_type": "send_money",
-		"from":       secondMemberRef,
-		"to":         firstMemberRef,
-		"amount":     100,
-	})
-	unmarshalResponse(t, secondBody, transferResponse)
-	assert.Equal(t, true, transferResponse.Success)
-
-	newFirstBalance := getBalance(t, firstMemberRef)
-	newSecondBalance := getBalance(t, secondMemberRef)
-
-	assert.Equal(t, oldFirstBalance+200, newFirstBalance)
-	assert.Equal(t, oldSecondBalance-200, newSecondBalance)
-}
-
-func TestCreateMembersWithSameName(t *testing.T) {
-	body := getResponseBody(t, postParams{
-		"query_type": "create_member",
-		"name":       "NameForTestCreateMembersWithSameName",
-	})
-
-	memberResponse := &createMemberResponse{}
-	unmarshalResponse(t, body, memberResponse)
-
-	firstMemberRef := memberResponse.Reference
-	assert.NotEqual(t, "", firstMemberRef)
-
-	body = getResponseBody(t, postParams{
-		"query_type": "create_member",
-		"name":       "NameForTestCreateMembersWithSameName",
-	})
-
-	unmarshalResponse(t, body, memberResponse)
-
-	secondMemberRef := memberResponse.Reference
-	assert.NotEqual(t, "", secondMemberRef)
-
-	assert.NotEqual(t, firstMemberRef, secondMemberRef)
 }
